+++ src/main/java/com/ibm/restconf/model/MessageType.java
@@ -1,0 +1,12 @@
+package com.ibm.restconf.model;
+
+public enum MessageType {
+    REQUEST,
+    RESPONSE,
+    MESSAGE;
+
+    @Override
+    public String toString() {
+        return this.name().toLowerCase();
+    }
+}